removing provider and forker: it violates microservice architecture
